+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changelog
 - Allow setting a default zoom level for both UI and documents
 - Drag and drop files from native file manager into wournal to open
 - Tooltip for shortcuts in toolbar buttons (not just in menu)
+- List of recent files in menu and in taskbar
 
 #### Fixed